packages/connector/pyproject.toml
[project]
name = "connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matthias", email = "matthias@repath.earth" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.13",
    "prefect>=3.4.1",
    "prefect-aws>=0.5.10",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
